+++ src/main/java/chess/service/RoomService.java
@@ -3,7 +3,9 @@ package chess.service;
 import chess.dao.RoomDao;
 import chess.domain.GameStatus;
 import chess.domain.chesspiece.Color;
+import chess.dto.CurrentTurnDto;
 import chess.dto.RoomStatusDto;
+import chess.view.JsonGenerator;
 
 public class RoomService {
 
@@ -23,5 +25,17 @@ public class RoomService {
         if (dto.getGameStatus().isEnd()) {
             roomDao.delete(roomName);
         }
+    }
+
+    public JsonGenerator findCurrentTurn(final String roomName) {
+        final JsonGenerator result = JsonGenerator.create();
+        try {
+            final RoomDao roomDao = new RoomDao();
+            final CurrentTurnDto dto = roomDao.findCurrentTurnByName(roomName);
+            result.addCurrentTurn(dto.getCurrentTurn());
+        } catch (IllegalArgumentException e) {
+            result.addError(e.getMessage());
+        }
+        return result;
     }
 }
